[Refactor] Mail sender separation - async logic - email limitation logic for 1 minute
