+++ src/main/java/com/cpayusin/ServerApplication.java
@@ -3,9 +3,11 @@ package com.cpayusin;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @EnableScheduling
+@EnableAsync
 @EnableJpaAuditing
 @SpringBootApplication
 public class ServerApplication
