+++ src/main/java/com/cpayusin/common/service/RedisService.java
@@ -1,57 +1,14 @@
 package com.cpayusin.common.service;
 
-import com.cpayusin.common.security.jwt.JwtService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Repository;
-
-import java.time.Duration;
-
-@Slf4j
-@RequiredArgsConstructor
-@Repository
-public class RedisService
+public interface RedisService
 {
-    private final RedisTemplate<String, String> redisTemplate;
-    private final JwtService jwtService;
+    void saveRefreshToken(String refreshToken, String email);
+    void deleteEmailAfterVerification(String email);
+    void saveEmailAndVerificationCodeWith5Minutes(String email, String verificationCode);
 
-    public void saveRefreshToken(String refreshToken, String email)
-    {
-        int refreshTokenExpirationMinutes = jwtService.getRefreshTokenExpirationMinutes();
-
-        redisTemplate.opsForValue().set(refreshToken, email, Duration.ofMinutes(refreshTokenExpirationMinutes));
-    }
-
-    public void deleteEmailAfterVerification(String email)
-    {
-        redisTemplate.delete(email);
-    }
-
-    public void saveEmailAndVerificationCodeWith5Minutes(String email, String verificationCode)
-    {
-        redisTemplate.opsForValue().set(email, verificationCode, Duration.ofMinutes(5));
-    }
-
-    public String getVerificationCodeByEmail(String email)
-    {
-        return redisTemplate.opsForValue().get(email);
-    }
-
-    public void deleteRefreshToken(String refreshToken)
-    {
-        redisTemplate.delete(refreshToken);
-    }
-
-    public Boolean hasKey(String refreshToken)
-    {
-        try{
-            jwtService.isValidToken(refreshToken);
-        }catch (RuntimeException ex){
-            log.error(ex.getMessage(), ex);
-            return false;
-        }
-
-        return redisTemplate.hasKey(refreshToken);
-    }
+    void saveEmailForLimitationFor1Minute(String email);
+    boolean isEmailLimited(String email);
+    String getVerificationCodeByEmail(String email);
+    void deleteRefreshToken(String refreshToken);
+    Boolean hasKey(String refreshToken);
 }
