+++ src/main/java/com/cpayusin/member/service/MailServiceImpl.java
@@ -1,19 +1,12 @@
 package com.cpayusin.member.service;
 
-import com.cpayusin.common.exception.BusinessLogicException;
-import com.cpayusin.common.exception.ExceptionMessage;
-import com.cpayusin.member.controller.request.SendVerificationCodeRequest;
 import com.cpayusin.common.service.RedisService;
+import com.cpayusin.member.service.port.MailSendHelper;
 import com.cpayusin.member.controller.port.MailService;
-import jakarta.mail.MessagingException;
-import jakarta.mail.internet.MimeMessage;
+import com.cpayusin.member.controller.request.SendVerificationCodeRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Service;
-import org.thymeleaf.context.Context;
-import org.thymeleaf.spring6.SpringTemplateEngine;
 
 import static com.cpayusin.common.service.UtilService.generateVerificationCode;
 
@@ -22,8 +15,7 @@ import static com.cpayusin.common.service.UtilService.generateVerificationCode;
 @Service
 public class MailServiceImpl implements MailService
 {
-    private final SpringTemplateEngine templateEngine;
-    private final JavaMailSender mailSender;
+    private final MailSendHelper helper;
     private final RedisService redisService;
 
     public String sendVerificationCode(SendVerificationCodeRequest request)
@@ -31,34 +23,16 @@ public class MailServiceImpl implements MailService
         String email = request.getEmail();
 
         String verificationCode = generateVerificationCode();
-        redisService.saveEmailAndVerificationCodeWith5Minutes(email, verificationCode);
-        return sendVerificationEmail(email, verificationCode);
-    }
 
-    private String sendVerificationEmail(String email, String verificationCode)
-    {
-        try {
-            MimeMessage mimeMessage = createMimeMessage(email, verificationCode);
-            mailSender.send(mimeMessage);
-            return "인증코드가 발송되었습니다. 5분 내로 인증을 완료해주세요.";
-        } catch (MessagingException e) {
-            log.error("Error sending verification email to {}", email, e);
-            throw new BusinessLogicException(ExceptionMessage.MAIL_ERROR);
+        if(!redisService.isEmailLimited(email)){
+            return "잠시 후 다시 시도해주세요.";
         }
+
+        redisService.saveEmailAndVerificationCodeWith5Minutes(email, verificationCode);
+        redisService.saveEmailForLimitationFor1Minute(email);
+
+        helper.sendVerificationEmail(email, verificationCode);
+        return "인증코드가 발송되었습니다. 5분 내로 인증을 완료해주세요.";
     }
 
-    private MimeMessage createMimeMessage(String email, String verificationCode) throws MessagingException
-    {
-        Context context = new Context();
-        context.setVariable("verificationCode", verificationCode);
-        String htmlContent = templateEngine.process("mailTemplate", context);
-
-        MimeMessage mimeMessage = mailSender.createMimeMessage();
-        MimeMessageHelper helper = new MimeMessageHelper(mimeMessage, "UTF-8");
-        helper.setTo(email);
-        helper.setSubject("이메일 인증 코드");
-        helper.setText(htmlContent, true);
-
-        return mimeMessage;
-    }
 }
