+++ src/main/java/com/cpayusin/member/service/port/MailSendHelper.java
@@ -1,0 +1,6 @@
+package com.cpayusin.member.service.port;
+
+public interface MailSendHelper
+{
+    void sendVerificationEmail(String email, String verificationCode);
+}
